Developer must call `require 'jsonapi/resources/matchers'`
When installing directly in a Rails application, adding `require: false`
in the Gemfile was sufficient. However, when using
jsonapi-resources-matchers in a Rails engine, it would load the gem
in the Rails app and cause issues.

To simplify things, and not need `require: false` in the Gemfile,
use must require the matchers file instead.